requirements.txt
pandas
numpy
xgboost
scikit-learn
joblib
apache-airflow==2.9.0
matplotlib
seaborn
